+++ java/lib/src/main/java/com/svix/models/OperationalWebhookEndpointOut.java
@@ -25,7 +25,7 @@ import java.util.Set;
 public class OperationalWebhookEndpointOut {
     @JsonProperty private String id;
     @JsonProperty private String description;
-    @JsonProperty private Long throttleRate;
+    @JsonProperty private Short throttleRate;
     @JsonProperty private String uid;
     @JsonProperty private URI url;
     @JsonProperty private Boolean disabled;
@@ -74,7 +74,7 @@ public class OperationalWebhookEndpointOut {
         this.description = description;
     }
 
-    public OperationalWebhookEndpointOut throttleRate(Long throttleRate) {
+    public OperationalWebhookEndpointOut throttleRate(Short throttleRate) {
         this.throttleRate = throttleRate;
         return this;
     }
@@ -87,11 +87,11 @@ public class OperationalWebhookEndpointOut {
      * @return throttleRate
      */
     @javax.annotation.Nullable
-    public Long getThrottleRate() {
+    public Short getThrottleRate() {
         return throttleRate;
     }
 
-    public void setThrottleRate(Long throttleRate) {
+    public void setThrottleRate(Short throttleRate) {
         this.throttleRate = throttleRate;
     }
