+++ java/lib/src/main/java/com/svix/models/SinkInCommon.java
@@ -22,7 +22,7 @@ import java.util.Set;
 @JsonAutoDetect(getterVisibility = Visibility.NONE, setterVisibility = Visibility.NONE)
 public class SinkInCommon {
     @JsonProperty private String description;
-    @JsonProperty private Long throttleRate;
+    @JsonProperty private Short throttleRate;
     @JsonProperty private String uid;
     @JsonProperty private String secret;
     @JsonProperty private Boolean disabled;
@@ -51,7 +51,7 @@ public class SinkInCommon {
         this.description = description;
     }
 
-    public SinkInCommon throttleRate(Long throttleRate) {
+    public SinkInCommon throttleRate(Short throttleRate) {
         this.throttleRate = throttleRate;
         return this;
     }
@@ -64,11 +64,11 @@ public class SinkInCommon {
      * @return throttleRate
      */
     @javax.annotation.Nullable
-    public Long getThrottleRate() {
+    public Short getThrottleRate() {
         return throttleRate;
     }
 
-    public void setThrottleRate(Long throttleRate) {
+    public void setThrottleRate(Short throttleRate) {
         this.throttleRate = throttleRate;
     }
