+++ rust/src/models/ingest_source_consumer_portal_access_in.rs
@@ -7,7 +7,7 @@ pub struct IngestSourceConsumerPortalAccessIn {
     ///
     /// Valid values are between 1 hour and 7 days. The default is 7 days.
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub expiry: Option<i32>,
+    pub expiry: Option<u64>,
 
     /// Whether the app portal should be in read-only mode.
     #[serde(rename = "readOnly")]
