+++ svix-cli/src/cmds/api/event_type.rs
@@ -8,7 +8,7 @@ use svix::models::*;
 pub struct EventTypeListOptions {
     /// Limit the number of returned items
     #[arg(long)]
-    pub limit: Option<i32>,
+    pub limit: Option<u64>,
     /// The iterator returned from a prior invocation
     #[arg(long)]
     pub iterator: Option<String>,
